requirements.txt
decorator==5.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.3
pillow==10.4.0
proglog==0.1.10
python-dotenv==1.0.1
tqdm==4.67.1
yt-dlp==2025.2.19
